requirements.txt
chromadb
sentence-transformers
streamlit
pypdf
python-dotenv
python-docx
openai